+++ CHANGELOG.md
@@ -16,6 +16,7 @@ IMPROVEMENT
 
 - [tendermint] Update tendermint to 0.23.0
 - [api] Add block reward to api
+- [api] Add bip_value field to Stake
 
 ## 0.1.9
 *Aug 19th, 2018*